[Chore] 주석 추가 및 클래스명 변경
