+++ app/src/main/java/kr/co/lion/modigm/db/study/StudyDao.kt
@@ -10,7 +10,11 @@ class StudyDao {
 
     private val tag = "StudyDao"
 
-    // 스터디와 스터디 멤버 데이터 조회 (좋아요 여부 포함)
+    /**
+     * 모든 스터디와 스터디 멤버 데이터 조회 (좋아요 여부 포함)
+     * @param userIdx 사용자 인덱스
+     * @return Result<List<Triple<SqlStudyData, Int, Boolean>>> 조회된 스터디 데이터를 반환
+     */
     suspend fun selectAllStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> =
         withContext(Dispatchers.IO) {
             runCatching {
@@ -43,8 +47,11 @@ class StudyDao {
             }
         }
 
-
-    // 특정 userIdx에 해당하는 스터디와 스터디 멤버 데이터 조회 (좋아요 여부 포함)
+    /**
+     * 특정 userIdx에 해당하는 스터디와 스터디 멤버 데이터 조회 (좋아요 여부 포함)
+     * @param userIdx 사용자 인덱스
+     * @return Result<List<Triple<SqlStudyData, Int, Boolean>>> 조회된 스터디 데이터를 반환
+     */
     suspend fun selectMyStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> =
         withContext(Dispatchers.IO) {
             runCatching {
@@ -78,7 +85,11 @@ class StudyDao {
             }
         }
 
-    // 좋아요한 스터디 목록 조회
+    /**
+     * 좋아요한 스터디 목록 조회
+     * @param userIdx 사용자 인덱스
+     * @return Result<List<Triple<SqlStudyData, Int, Boolean>>> 조회된 스터디 데이터를 반환
+     */
     suspend fun selectFavoriteStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> =
         withContext(Dispatchers.IO) {
             runCatching {
@@ -111,7 +122,12 @@ class StudyDao {
             }
         }
 
-    // 좋아요 추가 메소드
+    /**
+     * 좋아요 추가 메소드
+     * @param userIdx 사용자 인덱스
+     * @param studyIdx 스터디 인덱스
+     * @return Result<Boolean> 좋아요 추가 성공 여부를 반환
+     */
     suspend fun addFavorite(userIdx: Int, studyIdx: Int): Result<Boolean> =
         withContext(Dispatchers.IO) {
             runCatching {
@@ -130,7 +146,12 @@ class StudyDao {
             }
         }
 
-    // 좋아요 삭제 메소드
+    /**
+     * 좋아요 삭제 메소드
+     * @param userIdx 사용자 인덱스
+     * @param studyIdx 스터디 인덱스
+     * @return Result<Boolean> 좋아요 삭제 성공 여부를 반환
+     */
     suspend fun removeFavorite(userIdx: Int, studyIdx: Int): Result<Boolean> =
         withContext(Dispatchers.IO) {
             runCatching {
