+++ app/src/main/java/kr/co/lion/modigm/db/study/StudyDataSource.kt
@@ -9,7 +9,11 @@ class StudyDataSource {
     private val tag = "StudyDataSource"
     private val dao = StudyDao()
 
-    // 모집 중인 전체 스터디 목록 조회 (좋아요 여부 포함)
+    /**
+     * 모집 중인 전체 스터디 목록 조회 (좋아요 여부 포함)
+     * @param userIdx 사용자 인덱스
+     * @return Result<List<Triple<SqlStudyData, Int, Boolean>>> 조회된 스터디 데이터를 반환
+     */
     suspend fun getAllStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> {
         return runCatching {
             dao.selectAllStudyData(userIdx).getOrThrow()
@@ -19,16 +23,25 @@ class StudyDataSource {
         }
     }
 
-    // 내가 속한 스터디 목록 조회 (좋아요 여부 포함)
+    /**
+     * 내가 속한 스터디 목록 조회 (좋아요 여부 포함)
+     * @param userIdx 사용자 인덱스
+     * @return Result<List<Triple<SqlStudyData, Int, Boolean>>> 조회된 스터디 데이터를 반환
+     */
     suspend fun getMyStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> {
         return runCatching {
             dao.selectMyStudyData(userIdx).getOrThrow()
         }.onFailure { e ->
-            Log.e(tag, "전체 스터디 목록 조회 중 오류 발생", e)
+            Log.e(tag, "내 스터디 목록 조회 중 오류 발생", e)
             Result.failure<List<Triple<SqlStudyData, Int, Boolean>>>(e)
         }
     }
 
+    /**
+     * 좋아요한 스터디 목록 조회
+     * @param userIdx 사용자 인덱스
+     * @return Result<List<Triple<SqlStudyData, Int, Boolean>>> 조회된 스터디 데이터를 반환
+     */
     suspend fun getFavoriteStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> {
         return runCatching {
             dao.selectFavoriteStudyData(userIdx).getOrThrow()
@@ -38,7 +51,12 @@ class StudyDataSource {
         }
     }
 
-    // 좋아요 추가 메소드
+    /**
+     * 좋아요 추가 메소드
+     * @param userIdx 사용자 인덱스
+     * @param studyIdx 스터디 인덱스
+     * @return Result<Boolean> 좋아요 추가 성공 여부를 반환
+     */
     suspend fun addFavorite(userIdx: Int, studyIdx: Int): Result<Boolean> {
         return runCatching {
             dao.addFavorite(userIdx, studyIdx).getOrThrow()
@@ -48,7 +66,12 @@ class StudyDataSource {
         }
     }
 
-    // 좋아요 삭제 메소드
+    /**
+     * 좋아요 삭제 메소드
+     * @param userIdx 사용자 인덱스
+     * @param studyIdx 스터디 인덱스
+     * @return Result<Boolean> 좋아요 삭제 성공 여부를 반환
+     */
     suspend fun removeFavorite(userIdx: Int, studyIdx: Int): Result<Boolean> {
         return runCatching {
             dao.removeFavorite(userIdx, studyIdx).getOrThrow()
@@ -58,7 +81,9 @@ class StudyDataSource {
         }
     }
 
-    // 리소스를 해제하는 메서드 추가
+    /**
+     * 리소스를 해제하는 메서드
+     */
     suspend fun closeDataSource() {
         HikariCPDataSource.closeDataSource()
     }
